template deploy.yaml for more flexible secret handling
the fencing configuration requires an authentication key, but it's
deeply nested in the configuration data.  we could move the entire
fencing configuration into credentials.yaml, but it's nice to have it
part of the public repository.

we now generate deploy.yaml.in -> deploy.yaml using ansible-playbook
so that we can insert the authentication key into the configuration
file at deploy time.